Add GPLv3 license.
I originally made this as a quick throw together or proof of concept
thing so i didn't think to add a license but i guess people are
interested enough in this that it might matter (Issue #3).  GPL is
chosen since i'd rather this not be built in to some closed source
software, as much as that would even protect it from that but whichever.